% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-parsed-model.R
\name{as_parsed_model}
\alias{as_parsed_model}
\title{Prepares parsed model object}
\usage{
as_parsed_model(x)
}
\arguments{
\item{x}{A parsed model object}
}
\value{
The parsed model with its \code{parsed_model} and \verb{pm_*} classes set.
}
\description{
Prepares parsed model object
}
\examples{
pm <- parse_model(lm(mpg ~ wt, data = mtcars))
as_parsed_model(pm)

}
