% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-nested.R
\name{build_nested_node_stop}
\alias{build_nested_node_stop}
\title{Build nested node with "stop" NA handling}
\usage{
build_nested_node_stop(
  node_idx,
  tree_info,
  split_info,
  left_subtree,
  right_subtree
)
}
\description{
When the split variable is NA, predict at this node instead of continuing.
}
\keyword{internal}
