% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-nested.R
\name{build_nested_node_surrogate}
\alias{build_nested_node_surrogate}
\title{Build nested node with "surrogate" NA handling}
\usage{
build_nested_node_surrogate(
  node_idx,
  tree_info,
  split_info,
  left_subtree,
  right_subtree
)
}
\description{
Use surrogate splits to handle NA values, with majority direction as fallback.
}
\keyword{internal}
