% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-nested.R
\name{build_nested_split_condition}
\alias{build_nested_split_condition}
\title{Build a split condition expression for nested trees}
\usage{
build_nested_split_condition(split)
}
\arguments{
\item{split}{A split info list with col, val/vals, is_categorical}
}
\description{
Build a split condition expression for nested trees
}
\keyword{internal}
