% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-extractors.R
\name{deprecated-extractors}
\alias{deprecated-extractors}
\alias{.extract_xgb_trees}
\alias{.extract_lgb_trees}
\alias{.extract_catboost_trees}
\alias{.extract_rf_trees}
\alias{.extract_ranger_trees}
\alias{.extract_rf_classprob}
\alias{.extract_ranger_classprob}
\alias{.extract_rpart_classprob}
\alias{.extract_partykit_classprob}
\alias{.extract_earth_multiclass}
\alias{.extract_glmnet_multiclass}
\title{Deprecated model extractors}
\usage{
.extract_xgb_trees(model)

.extract_lgb_trees(model)

.extract_catboost_trees(model)

.extract_rf_trees(model)

.extract_ranger_trees(model)

.extract_rf_classprob(model)

.extract_ranger_classprob(model)

.extract_rpart_classprob(model)

.extract_partykit_classprob(model)

.extract_earth_multiclass(model)

.extract_glmnet_multiclass(model, penalty = NULL)
}
\arguments{
\item{model}{A fitted model object.}

\item{penalty}{The penalty value to use for coefficient extraction.}
}
\value{
The same values these functions have always returned. Note that
\code{\link[=tidypredict_class_exprs]{tidypredict_class_exprs()}} returns language objects where
\code{.extract_earth_multiclass()} and \code{.extract_glmnet_multiclass()} return
strings.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been replaced by generics with documented return
shapes. See \link{tidypredict_extractors}.\tabular{ll}{
   Deprecated \tab Replacement \cr
   \code{.extract_xgb_trees()}, \code{.extract_lgb_trees()}, \code{.extract_catboost_trees()}, \code{.extract_rf_trees()}, \code{.extract_ranger_trees()} \tab \code{\link[=tidypredict_trees]{tidypredict_trees()}} \cr
   \code{.extract_rf_classprob()}, \code{.extract_ranger_classprob()} \tab \code{\link[=tidypredict_class_trees]{tidypredict_class_trees()}} \cr
   \code{.extract_rpart_classprob()}, \code{.extract_partykit_classprob()}, \code{.extract_earth_multiclass()}, \code{.extract_glmnet_multiclass()} \tab \code{\link[=tidypredict_class_exprs]{tidypredict_class_exprs()}} \cr
}
}
\keyword{internal}
