% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-C5.0.R
\name{.c50_tree_info_full}
\alias{.c50_tree_info_full}
\title{Extract comprehensive tree info for C5.0 models}
\usage{
.c50_tree_info_full(model)
}
\arguments{
\item{model}{A C5.0 model object}
}
\description{
Returns tree structure in format needed by nested case_when generator.
For use in orbital package.
}
\keyword{internal}
