% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bagger.R
\name{.extract_bagger_classprob}
\alias{.extract_bagger_classprob}
\title{Extract class probability trees for bagger models}
\usage{
.extract_bagger_classprob(model)
}
\arguments{
\item{model}{A bagger model object (classification)}
}
\description{
Returns one list of per-tree expressions for each outcome level. For use in
orbital package.
}
\keyword{internal}
