% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bagger.R
\name{.extract_bagger_trees}
\alias{.extract_bagger_trees}
\title{Extract regression trees for bagger models}
\usage{
.extract_bagger_trees(model)
}
\arguments{
\item{model}{A bagger model object (regression)}
}
\description{
For use in orbital package.
}
\keyword{internal}
