% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bart.R
\name{.extract_bart_trees}
\alias{.extract_bart_trees}
\title{Extract the trees of a bart model}
\usage{
.extract_bart_trees(model)
}
\arguments{
\item{model}{A \code{dbarts::bart()} model object}
}
\description{
For use in orbital package.
}
\keyword{internal}
