% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-catboost.R
\name{.extract_catboost_trees}
\alias{.extract_catboost_trees}
\title{Extract processed CatBoost trees}
\usage{
.extract_catboost_trees(model)
}
\arguments{
\item{model}{A CatBoost model object}
}
\description{
For use in orbital package.
}
\keyword{internal}
