% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-earth.R
\name{.extract_earth_multiclass}
\alias{.extract_earth_multiclass}
\title{Extract multiclass linear predictors for earth models}
\usage{
.extract_earth_multiclass(model)
}
\arguments{
\item{model}{An earth model object with multiple classes (glm.list with >1 elements)}
}
\description{
For use in orbital package.
}
\keyword{internal}
