% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-glmnet.R
\name{.extract_glmnet_multiclass}
\alias{.extract_glmnet_multiclass}
\title{Extract multiclass linear predictors for glmnet models}
\usage{
.extract_glmnet_multiclass(model, penalty = NULL)
}
\arguments{
\item{model}{A glmnet model object with class "multnet"}

\item{penalty}{The penalty value to use for coefficient extraction}
}
\description{
For use in orbital package.
}
\keyword{internal}
