% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lightgbm.R
\name{.extract_lgb_trees}
\alias{.extract_lgb_trees}
\title{Extract processed LightGBM trees}
\usage{
.extract_lgb_trees(model)
}
\arguments{
\item{model}{A LightGBM model object}
}
\description{
For use in orbital package.
}
\keyword{internal}
