% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ranger.R
\name{.extract_ranger_classprob}
\alias{.extract_ranger_classprob}
\title{Extract classification probability trees for ranger models}
\usage{
.extract_ranger_classprob(model)
}
\arguments{
\item{model}{A ranger model object fitted with \code{probability = TRUE}}
}
\description{
For use in orbital package.
}
\keyword{internal}
