% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ranger.R
\name{.extract_ranger_trees}
\alias{.extract_ranger_trees}
\title{Extract regression trees for ranger models}
\usage{
.extract_ranger_trees(model)
}
\arguments{
\item{model}{A ranger model object (regression)}
}
\description{
For use in orbital package.
}
\keyword{internal}
