% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-rf.R
\name{.extract_rf_classprob}
\alias{.extract_rf_classprob}
\title{Extract classification vote trees for randomForest models}
\usage{
.extract_rf_classprob(model)
}
\arguments{
\item{model}{A randomForest model object}
}
\description{
For use in orbital package.
}
\keyword{internal}
