% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-rf.R
\name{.extract_rf_trees}
\alias{.extract_rf_trees}
\title{Extract regression trees for randomForest models}
\usage{
.extract_rf_trees(model)
}
\arguments{
\item{model}{A randomForest model object (regression)}
}
\description{
For use in orbital package.
}
\keyword{internal}
