% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-rpart.R
\name{.extract_rpart_classprob}
\alias{.extract_rpart_classprob}
\title{Extract classprob trees for rpart models}
\usage{
.extract_rpart_classprob(model, nested = FALSE, na_handling = "none")
}
\arguments{
\item{model}{An rpart model object}

\item{nested}{Logical, whether to use nested case_when (default FALSE)}

\item{na_handling}{How to handle NA values when nested = TRUE}
}
\description{
For use in orbital package.
}
\keyword{internal}
