% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-xgboost.R
\name{.extract_xgb_trees}
\alias{.extract_xgb_trees}
\title{Extract processed xgboost trees}
\usage{
.extract_xgb_trees(model)
}
\arguments{
\item{model}{An xgb.Booster model}
}
\description{
For use in orbital package.
}
\keyword{internal}
