% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-partykit.R
\name{.partykit_tree_info_full}
\alias{.partykit_tree_info_full}
\title{Extract comprehensive tree info for partykit models}
\usage{
.partykit_tree_info_full(model)
}
\arguments{
\item{model}{A partykit model object}
}
\description{
Returns tree structure in format needed by nested case_when generator.
For use in orbital package.
}
\keyword{internal}
