% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-rpart.R
\name{.rpart_tree_info_full}
\alias{.rpart_tree_info_full}
\title{Extract comprehensive tree info for rpart models}
\usage{
.rpart_tree_info_full(model)
}
\arguments{
\item{model}{An rpart model object}
}
\description{
Returns tree structure in format needed by nested case_when generator.
For use in orbital package.
}
\keyword{internal}
