% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql.R
\name{fit_to_sql}
\alias{fit_to_sql}
\title{Returns a SQL query with formula to calculate fitted values}
\usage{
fit_to_sql(model, con)
}
\arguments{
\item{model}{An R model or a tibble with a parsed model}

\item{con}{Database connection object. It is used to select 
the correst SQL translation syntax.}
}
\description{
Returns a SQL query with formula to calculate fitted values
}
\examples{

model <- lm(mpg ~ wt + am + cyl, data = mtcars)
fit_to_sql(model, dbplyr::simulate_dbi())

}
