% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql.R
\name{interval_to_sql}
\alias{interval_to_sql}
\title{Returns a SQL query with formula to calculate predicted interval}
\usage{
interval_to_sql(model, con, interval = 0.95)
}
\arguments{
\item{model}{An R model or a tibble with a parsed model}

\item{con}{Database connection object. It is used to select 
the correst SQL translation syntax.}

\item{interval}{The prediction interval, defaults to 0.95}
}
\description{
Returns a SQL query with formula to calculate predicted interval
}
\examples{

model <- lm(mpg ~ wt + am + cyl, data = mtcars)
interval_to_sql(model, dbplyr::simulate_dbi())

}
