% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-predictions.R
\name{knit_print.test_predictions}
\alias{knit_print.test_predictions}
\title{Knit print method for test predictions results}
\usage{
knit_print.test_predictions(x, ...)
}
\description{
Knit print method for test predictions results
}
\keyword{internal}
