% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-predictions.R
\name{knit_print.tidypredict_test}
\alias{knit_print.tidypredict_test}
\title{Knit print method for test predictions results}
\usage{
\method{knit_print}{tidypredict_test}(x, ...)
}
\description{
Knit print method for test predictions results
}
\keyword{internal}
