% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsemodel.R
\name{parsemodel}
\alias{parsemodel}
\title{Converts an R model object into a table}
\usage{
parsemodel(model)
}
\arguments{
\item{model}{An R model object}
}
\description{
It parses a fitted R model's structure to extract all of the necessary
components to pass along to the functions that create the actual 
Tidy Eval formula that dplyr can read.  It is currently able to parse
lm() and glm() models.
}
\examples{

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
parsemodel(model)

}
