% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{path_formulas}
\alias{path_formulas}
\title{Turn a path object into a combined expression}
\usage{
path_formulas(path)
}
\arguments{
\item{path}{a list of lists.

This list can contain 0 or more elemements. The elements but each be of the
following format:
\itemize{
\item \code{type} character, must be \code{"conditional"}, \code{"set"}, or \code{"all"}.
\item \code{op} character.
if \code{type == "conditional"} must be \code{"more"}, \code{"more-equal"}, \code{"less"}, or
\code{"less-equal"}.
if \code{type == "set"} must be \code{"in"} on \verb{not-in}.
\item \code{col} character.
\item \code{val} if \code{type == "conditional"} and \code{vals} if \code{type == "set"}.
Can be character or numeric.
}}
}
\description{
Turn a path object into a combined expression
}
\keyword{internal}
