% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypredict.R
\name{predict_interval}
\alias{predict_interval}
\title{Returns a Tidy Eval formula to calculate prediction interval}
\usage{
predict_interval(model, interval = 0.95)
}
\arguments{
\item{model}{An R model object}

\item{interval}{The prediction interval, defaults to 0.95}
}
\description{
It uses parsemodel to create a tabular version of the data needed to create
a Tidy Eval formula that can then be used inside a dplyr command.  It uses 
S3 methods to automatically determine the class of the model and run the
corresponding function that is able to create the Tidy Eval formula.  It
currently supports lm() models.
}
\details{
The result still has to be added to the fit to obtain the upper bound, and
substracted from fit to obtain the lower bound.
}
\examples{

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
predict_interval(model)


}
