% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-column.R
\name{predict_to_column}
\alias{predict_to_column}
\title{Adds the prediction columns to a piped command set}
\usage{
predict_to_column(df, model, add_interval = FALSE, interval = 0.95,
  vars = c("fit", "upper", "lower"))
}
\arguments{
\item{df}{A data.frame or tibble}

\item{model}{An R object, typically an R model, but it can also bee a regular data frame 
that has a parsed model already loaded.}

\item{add_interval}{Switch that indicates if the prediction interval columns should be added. Defaults
to FALSE}

\item{interval}{The prediction interval, defaults to 0.95. It is ignored if add_interval is set to
FALSE}

\item{vars}{The name of the variables that this function will produce. It defaults to "fit", "upper", and "lower".}
}
\description{
Adds a new column with the results form predict_fit() to a piped command set.
If add_interval is set to TRUE, then it will add two additional columns, one 
for the lower and another for the upper prediction interval bounds.
}
\examples{

library(dplyr)

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
df \%>\%
  predict_to_column(model, add_interval = TRUE)

}
