% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-predictions.R
\name{print.test_predictions}
\alias{print.test_predictions}
\title{print method for test predictions results}
\usage{
\method{print}{test_predictions}(x, ...)
}
\description{
print method for test predictions results
}
\keyword{internal}
