% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypredict_test.R
\name{print.tidypredict_test}
\alias{print.tidypredict_test}
\title{print method for test predictions results}
\usage{
\method{print}{tidypredict_test}(x, ...)
}
\description{
print method for test predictions results
}
\keyword{internal}
