% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-predictions.R
\name{test_predictions}
\alias{test_predictions}
\title{Tests base predict function against tidypredict}
\usage{
test_predictions(model, df = model$model, threshold = 1e-12,
  include_intervals = FALSE, max_rows = NULL)
}
\arguments{
\item{model}{An R model object}

\item{df}{A data frame that contains all of the needed fields to run the prediction.
It defaults to the "model" data frame object inside the model object.}

\item{threshold}{The number that a given result difference, between predict() and
predict_to_column() should not exceed.  For continuous predictions, the default 
value is 0.000000000001 (1e-12), for categorical predictions, the default value is
0.}

\item{include_intervals}{Switch to indicate if the prediction intervals should be
included in the test.  It defaults to FALSE.}

\item{max_rows}{The number of rows in the object passed in the df argument. Highly
recommended for large data sets.}
}
\description{
Compares the results of running predict() and the predict_to_column() 
functions.
}
\examples{

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
test_predictions(model, include_intervals = TRUE)

}
