% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels.R
\name{tidy.pm_regression}
\alias{tidy.pm_regression}
\title{Tidy the parsed model results}
\usage{
\method{tidy}{pm_regression}(x, ...)
}
\arguments{
\item{x}{A parsed_model object}

\item{...}{Reserved for future use}
}
\value{
A tibble with one row per term, containing the \code{term} name and its
\code{estimate}.
}
\description{
Tidy the parsed model results
}
\examples{
pm <- parse_model(lm(mpg ~ wt, data = mtcars))
tidy(pm)

}
