% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-fit.R
\name{tidypredict_fit}
\alias{tidypredict_fit}
\title{Returns a Tidy Eval formula to calculate fitted values}
\usage{
tidypredict_fit(model)
}
\arguments{
\item{model}{An R model or a list with a parsed model.}
}
\value{
A Tidy Eval formula that calculates the fitted values. Models that
produce one formula per class or per outcome return a list of formulas.
}
\description{
It parses a model or uses an already parsed model to return a
Tidy Eval formula that can then be used inside a dplyr command.
}
\examples{

model <- lm(mpg ~ wt + cyl * disp, offset = am, data = mtcars)
tidypredict_fit(model)
}
