% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-interval.R
\name{tidypredict_interval}
\alias{tidypredict_interval}
\title{Returns a Tidy Eval formula to calculate prediction interval.}
\usage{
tidypredict_interval(model, interval = 0.95)
}
\arguments{
\item{model}{An R model or a list with a parsed model}

\item{interval}{The prediction interval, defaults to 0.95}
}
\value{
A Tidy Eval formula that calculates the half width of the
prediction interval. It must be added to and subtracted from the fit to
obtain the upper and lower bounds.
}
\description{
It parses a model or uses an already parsed model to return a
Tidy Eval formula that can then be used inside a dplyr command.
}
\details{
The result still has to be added to and subtracted from the fit to obtain the upper and
lower bound respectively.
}
\examples{

model <- lm(mpg ~ wt + cyl * disp, offset = am, data = mtcars)
tidypredict_interval(model)
}
