% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql.R
\name{tidypredict_sql}
\alias{tidypredict_sql}
\title{Returns a SQL query with formula to calculate fitted values}
\usage{
tidypredict_sql(model, con)
}
\arguments{
\item{model}{An R model or a list with a parsed model}

\item{con}{Database connection object. It is used to select
the correct SQL translation syntax.}
}
\value{
A SQL query, as returned by \code{\link[dbplyr:translate_sql]{dbplyr::translate_sql()}}. Models that
produce one formula per class or per outcome return a list of queries.
}
\description{
Returns a SQL query with formula to calculate fitted values
}
\examples{
\dontshow{if (rlang::is_installed("dbplyr")) withAutoprint(\{ # examplesIf}
model <- lm(mpg ~ wt + am + cyl, data = mtcars)
tidypredict_sql(model, dbplyr::simulate_dbi())
\dontshow{\}) # examplesIf}
}
