% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-column.R
\name{tidypredict_to_column}
\alias{tidypredict_to_column}
\title{Adds the prediction columns to a piped command set.}
\usage{
tidypredict_to_column(
  df,
  model,
  add_interval = FALSE,
  interval = 0.95,
  vars = c("fit", "upper", "lower")
)
}
\arguments{
\item{df}{A data.frame or tibble}

\item{model}{An R model or a parsed model inside a data frame}

\item{add_interval}{Switch that indicates if the prediction interval columns should be added. Defaults
to FALSE}

\item{interval}{The prediction interval, defaults to 0.95. Ignored if add_interval is set to
FALSE}

\item{vars}{The name of the variables that this function will produce. Defaults to "fit", "upper", and "lower".}
}
\value{
The input data frame with one new column (the fit) added, or three
new columns (fit, upper and lower bounds) when \code{add_interval} is \code{TRUE}.
}
\description{
Adds a new column with the results from tidypredict_fit() to a piped command set.
If add_interval is set to TRUE, it will add two additional columns- one
for the lower and another for the upper prediction interval bounds.
}
\examples{
model <- lm(mpg ~ wt, data = mtcars)
tidypredict_to_column(mtcars, model)

}
