% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptable.R
\name{acceptable_formula}
\alias{acceptable_formula}
\title{Checks that the formula can be parsed}
\usage{
acceptable_formula(model)
}
\arguments{
\item{model}{An R model object}
}
\description{
Uses an S3 method to check that a given formula can be parsed based on its class.
Currently it supports lm() and glm() models.  It currently checks for contrasts
in variables that currently cannot be parsed, and checks to make sure that there
are no in-line R functions in the call of the formula (e.g: lm(wt ~ as.factor(am))).
This function is mostly meant for function interaction, not user interaction.
}
\examples{

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
acceptable_formula(model)

}
