% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-partykit.R
\name{.extract_partykit_classprob}
\alias{.extract_partykit_classprob}
\title{Extract classprob trees for partykit models}
\usage{
.extract_partykit_classprob(model)
}
\description{
For use in orbital package.
}
\keyword{internal}
