% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-rpart.R
\name{.extract_rpart_classprob}
\alias{.extract_rpart_classprob}
\title{Extract classprob trees for rpart models}
\usage{
.extract_rpart_classprob(model)
}
\arguments{
\item{model}{An rpart model object}
}
\description{
For use in orbital package.
}
\keyword{internal}
