% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-nested.R
\name{generate_nested_case_when_tree}
\alias{generate_nested_case_when_tree}
\title{Generate nested case_when for a tree}
\usage{
generate_nested_case_when_tree(tree_info)
}
\arguments{
\item{tree_info}{A tree info list from \code{rpart_tree_info_full()} or similar}
}
\description{
Generate nested case_when for a tree
}
\keyword{internal}
