% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{path_formula}
\alias{path_formula}
\title{Turn a path object into an expression}
\usage{
path_formula(x)
}
\arguments{
\item{x}{a list.

The input of this function is a list with 4 values.
\itemize{
\item \code{type} character, must be \code{"conditional"} or \code{"set"}.
\item \code{op} character.
if \code{type == "conditional"} must be \code{"more"}, \code{"more-equal"}, \code{"less"}, or
\code{"less-equal"}.
if \code{type == "set"} must be \code{"in"} on \verb{not-in}.
\item \code{col} character.
\item \code{val} if \code{type == "conditional"} and \code{vals} if \code{type == "set"}.
Can be character or numeric.
@keywords internal
}}
}
\description{
Turn a path object into an expression
}
