% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypredict-package.R
\docType{package}
\name{tidypredict-package}
\alias{tidypredict}
\alias{tidypredict-package}
\title{tidypredict: Run Predictions Inside the Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

It parses a fitted 'R' model object, and returns a formula in 'Tidy Eval' code that calculates the predictions. It works with several databases back-ends because it leverages 'dplyr' and 'dbplyr' for the final 'SQL' translation of the algorithm. It currently supports lm(), glm(), randomForest(), ranger(), rpart(), C5.0(), earth(), xgb.Booster.complete(), lgb.Booster(), catboost.Model(), cubist(), ctree(), cforest(), fda(), ksvm(), lda(), sda(), lda_diag(), and xrf() models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidypredict.tidymodels.org}
  \item \url{https://github.com/tidymodels/tidypredict}
  \item Report bugs at \url{https://github.com/tidymodels/tidypredict/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co}

Authors:
\itemize{
  \item Emil Hvitfeldt \email{emil.hvitfeldt@posit.co}
  \item Edgar Ruiz \email{edgar@posit.co}
  \item Max Kuhn \email{max@posit.co}
}

}
\keyword{internal}
