% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypredict.R
\docType{package}
\name{tidypredict-package}
\alias{tidypredict}
\alias{tidypredict-package}
\title{tidypredict: Run Predictions Inside the Database}
\description{
It parses a fitted 'R' model object, and returns a formula
in 'Tidy Eval' code that calculates the predictions.
It works with several databases back-ends because it leverages 'dplyr'
and 'dbplyr' for the final 'SQL' translation of the algorithm. It currently
supports lm(), glm() and randomForest() models.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://tidypredict.netlify.com/}
  \item Report bugs at \url{https://github.com/edgararuiz/tidypredict/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgar@rstudio.com}

}
\keyword{internal}
