% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-fit.R
\name{tidypredict_fit}
\alias{tidypredict_fit}
\title{Returns a Tidy Eval formula to calculate fitted values}
\usage{
tidypredict_fit(model)
}
\arguments{
\item{model}{An R model or a tibble with a parsed model}
}
\description{
It uses parse_model to create a tabular version of the data needed to create
a Tidy Eval formula that can then be used inside a dplyr command.  It uses
S3 methods to automatically determine the class of the model and run the
corresponding function that is able to create the Tidy Eval formula.  It
currently supports lm() and glm() models.
}
\examples{

df <- data.frame(x = c(1, 2, 5, 6 ,6), y = c(2, 3, 6, 5, 4))
model <- lm(x ~ y, df)
tidypredict_fit(model)


}
