% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql.R
\name{tidypredict_sql}
\alias{tidypredict_sql}
\title{Returns a SQL query with formula to calculate fitted values}
\usage{
tidypredict_sql(model, con)
}
\arguments{
\item{model}{An R model or a tibble with a parsed model}

\item{con}{Database connection object. It is used to select
the correct SQL translation syntax.}
}
\description{
Returns a SQL query with formula to calculate fitted values
}
\examples{
library(dbplyr)

model <- lm(mpg ~ wt + am + cyl, data = mtcars)
tidypredict_sql(model, simulate_dbi())

}
\keyword{internal}
